broker/requirements.txt
confluent-kafka==2.6.0
paho-mqtt==2.1.0
py4j==0.10.9.7
pyspark==3.5.3
